Moved timer source files
